some cleanup and improvements to Uno
