+++ src/skiasharp/LiveChartsCore.SkiaSharp.Xamarin.Forms/LiveChartsCore.SkiaSharpView.Xamarin.Forms.csproj
@@ -6,7 +6,7 @@
     <TargetFrameworks>netstandard2.0;</TargetFrameworks>
     <AssemblyName>LiveChartsCore.SkiaSharpView.XamarinForms</AssemblyName>
     <RootNamespace>LiveChartsCore.SkiaSharpView.XamarinForms</RootNamespace>
-    <Version>2.0.0-beta.210</Version>
+    <Version>2.0.0-beta.220</Version>
     <PackageIcon>icon.png</PackageIcon>
     <Description>Simple, flexible, interactive and powerful data visualization for XamarinForms.</Description>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
