+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Eto/LiveChartsCore.SkiaSharpView.Eto.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Library</OutputType>
@@ -7,7 +7,7 @@
     <TargetFrameworks>netstandard2.0</TargetFrameworks>
     <AssemblyName>LiveChartsCore.SkiaSharpView.Eto</AssemblyName>
     <RootNamespace>LiveChartsCore.SkiaSharpView.Eto</RootNamespace>
-    <Version>2.0.0-beta.210</Version>
+    <Version>2.0.0-beta.220</Version>
     <PackageIcon>icon.png</PackageIcon>
     <Description>Simple, flexible, interactive and powerful data visualization for Eto.Forms.</Description>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
