+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Maui/LiveChartsCore.SkiaSharpView.Maui.nuspec
@@ -2,7 +2,7 @@
 <package>
   <metadata>
     <id>LiveChartsCore.SkiaSharpView.Maui</id>
-    <version>2.0.0-beta.210</version>
+    <version>2.0.0-beta.220</version>
     <title>LiveChartsCore.SkiaSharpView.Maui</title>
     <authors>BetoRodriguez</authors>
     <requireLicenseAcceptance>true</requireLicenseAcceptance>
@@ -17,23 +17,23 @@
     <dependencies>
       <group targetFramework="net6.0">
         <dependency id="SkiaSharp.Views.Maui.Controls.Compatibility" version="2.88.0-preview.254" />
-        <dependency id="LiveChartsCore.SkiaSharpView" version="2.0.0-beta.210" />
+        <dependency id="LiveChartsCore.SkiaSharpView" version="2.0.0-beta.220" />
       </group>
       <group targetFramework="net6.0-ios13.6">
         <dependency id="SkiaSharp.Views.Maui.Controls.Compatibility" version="2.88.0-preview.254" />
-        <dependency id="LiveChartsCore.SkiaSharpView" version="2.0.0-beta.210" />
+        <dependency id="LiveChartsCore.SkiaSharpView" version="2.0.0-beta.220" />
       </group>
       <group targetFramework="net6.0-maccatalyst13.5">
         <dependency id="SkiaSharp.Views.Maui.Controls.Compatibility" version="2.88.0-preview.254" />
-        <dependency id="LiveChartsCore.SkiaSharpView" version="2.0.0-beta.210" />
+        <dependency id="LiveChartsCore.SkiaSharpView" version="2.0.0-beta.220" />
       </group>
       <group targetFramework="net6.0-android30.0">
         <dependency id="SkiaSharp.Views.Maui.Controls.Compatibility" version="2.88.0-preview.254" />
-        <dependency id="LiveChartsCore.SkiaSharpView" version="2.0.0-beta.210" />
+        <dependency id="LiveChartsCore.SkiaSharpView" version="2.0.0-beta.220" />
       </group>
       <group targetFramework="net6.0-windows10.0.18362">
         <dependency id="SkiaSharp.Views.Maui.Controls.Compatibility" version="2.88.0-preview.254" />
-        <dependency id="LiveChartsCore.SkiaSharpView" version="2.0.0-beta.210" />
+        <dependency id="LiveChartsCore.SkiaSharpView" version="2.0.0-beta.220" />
       </group>
     </dependencies>
   </metadata>
