+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Uno/CartesianChart.xaml.cs
@@ -1,4 +1,5 @@
-﻿// The MIT License(MIT)
+﻿
+// The MIT License(MIT)
 //
 // Copyright(c) 2021 Alberto Rodriguez Orozco & LiveCharts Contributors
 //
