+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Uno/DefaultLegend.xaml.cs
@@ -24,6 +24,7 @@ using System.Linq;
 using LiveChartsCore.Kernel.Sketches;
 using LiveChartsCore.Measure;
 using LiveChartsCore.SkiaSharpView.Drawing;
+using LiveChartsCore.SkiaSharpView.Uno.Binding;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
