+++ src/skiasharp/LiveChartsCore.SkiaSharpView.WinUI/LiveChartsCore.SkiaSharpView.WinUI.nuspec
@@ -2,7 +2,7 @@
 <package>
   <metadata>
     <id>LiveChartsCore.SkiaSharpView.WinUI</id>
-    <version>2.0.0-beta.210</version>
+    <version>2.0.0-beta.220</version>
     <title>LiveChartsCore.SkiaSharpView.WinUI</title>
     <authors>BetoRodriguez</authors>
     <requireLicenseAcceptance>true</requireLicenseAcceptance>
@@ -19,8 +19,8 @@
       <group targetFramework="net5.0-windows10.0.18362">
         <dependency id="Newtonsoft.Json" version="[13.0.1]" />
         <dependency id="SkiaSharp.Views.WinUI" version="2.88.0-preview.254" />
-        <dependency id="LiveChartsCore" version="2.0.0-beta.210" />
-        <dependency id="LiveChartsCore.SkiaSharpView" version="2.0.0-beta.210" />
+        <dependency id="LiveChartsCore" version="2.0.0-beta.220" />
+        <dependency id="LiveChartsCore.SkiaSharpView" version="2.0.0-beta.220" />
       </group>
   </dependencies>
   </metadata>
